+++ scripts/node/javet-rebuild/rebuild-sqlite3.sh
@@ -1,0 +1,2 @@
+export NODE_MODULE_FILE="../node_modules/sqlite3/lib/binding/napi-v3-linux-x64/node_sqlite3.node"
+./rebuild.sh