+++ src/test/java/com/caoccao/javet/interop/TestNodeRuntime.java
@@ -23,7 +23,6 @@ import com.caoccao.javet.exceptions.JavetException;
 import com.caoccao.javet.node.modules.NodeModuleAny;
 import com.caoccao.javet.node.modules.NodeModuleProcess;
 import com.caoccao.javet.utils.JavetOSUtils;
-import com.caoccao.javet.values.primitive.V8ValueString;
 import com.caoccao.javet.values.reference.V8ValueArray;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -53,7 +52,7 @@ public class TestNodeRuntime extends BaseTestJavet {
 
     @BeforeEach
     public void beforeEach() throws JavetException {
-        nodeRuntime = (NodeRuntime) v8Host.createV8Runtime();
+        nodeRuntime = v8Host.createV8Runtime();
         assertFalse(nodeRuntime.isPooled());
         assertEquals(0, nodeRuntime.getReferenceCount(),
                 "Reference count should be 0 before test case is started.");
@@ -102,6 +101,15 @@ public class TestNodeRuntime extends BaseTestJavet {
         Path path4 = nodeModuleProcess.getWorkingDirectory();
         assertNotEquals(path1.toAbsolutePath().toString(), path3.toAbsolutePath().toString());
         assertEquals(path1.toAbsolutePath().toString(), path4.toAbsolutePath().toString());
+    }
+
+    @Test
+    public void testSqlite3() throws JavetException {
+        File sqlite3File = getScriptFile("../node_modules/sqlite3/sqlite3.js");
+        if (sqlite3File.exists()) {
+            File scriptFile = getScriptFile("test-node-module-sqlite3-sync.js");
+            nodeRuntime.getExecutor(scriptFile).executeVoid();
+        }
     }
 
     @Test